requirements.txt
fastapi
uvicorn
python-multipart
groq
python-dotenv
psycopg2-binary
sqlachemy
pyjwt
bcrypt
